protocol: send-only rpc client
With the advent of the git-server, it is required to have a client
that can make RPC requests without being connected to the gossip
network. In fact, it only needs to know about the git protocol -- for
performing the replication response during a request-pull.

Provide an RPC client that only makes send requests, which include
`replicate`, `request-pull`, and `interrogation`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>

rpc::client: simplify by passing ingress

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>