peer: expose RPC methods via Client
The Peer and Client have similar functionality for the functions
`replicate`, `interrogate`, and `request_pull`.

We deprecate these functions and allow the use of them via
`Peer::client` -- which creates a `Client` based on the current
`Peer`.

The tests that use these functions are changed to use the `client` method.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>