Allow anyone to change a collaborative object
We effectively remove the authorization logic of `cobs` and accept any
change we are aware of. This implies that applications must perform
their own authorization logic. To allow this we extend the `cobs` API,
`cob::CollaborativeObject::history` becomes a representation of the
entire change graph - as opposed to just the authorized changes. Each
change in the graph is tagged with the URN of the identity which signed
the change. Applications can access this graph via
`cob::History::traverse`.

While we're here we also simplify the cache so that incompatible
changes in cache layout will just cause a reload of the cached objects.
This allows us to remove the fixture test for cache layout. We rename
`ThinChangeGraph` to `CachedChangeGraph` to better communicate it's use.

Signed-off-by: Alex Good <alex@memoryandthought.me>