nix: update niv sources to avoid lld warnings
When running cargo commands with rustc-1.60, the following warning
would be spit out to the terminal:

  rustc: /usr/local/lib/libz.so.1: no version information available

After a lot of digging and patchelf'ing, no solution was found. That
is, until `niv update` was run and everything worked fine again. No
attempt is made to understand why, but happiness is had now that the
terminal is not flooded with annoying messages.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>