gitd: wire up rpc-client
Teach the gitd server to use Client for making request-pull and
replicate calls in post-receive and pre-upload, respectively.

These only happen when the gitd server is configured with
`--push-seeds` and `--fetch-seeds`, to allow for use cases where one
does not want interact with the network upon git commands. This means
that the operations will only interact with the local storage.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>