lnk-clib: check if file store path is a directory
If the path provided to FileStore is a directory, the executing
program would begin to hog resources, with resident memory of the
process spiking into GB. This is due to `io::Lines` iterator
continuously returning an `io::Error` stating that the path is a
directory.

Defend against this by checking that path provided is a directory and
returning an error if it was and also make the iterator return None if
it has seen an `io::Error` from the `io::Lines` iterator.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>