lnk-sync: introduce lnk-sync crate
This functionality allows the synchronisation with the set of
configured seeds. It will replicate and request-pull for each seed
found. If the caller wants to only replicate or only request-pull, a
Mode is provided to achieve this.

Since the output is planned for the lnk CLI, they derive Serialize so
as the main call can print the output as JSON.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>