Add default_branch_head and set_default_branch
When checking out projects from the monorepo it is useful to set the
`refs/namespaces/<urn>/HEAD` reference to the default branch of the
project so that the resulting working copy is in a useful state (namely
pointing at the latest commit for the default branch).

In general this is not possible because delegates may have diverging
views of the project, but often they do not disagree. Add
`librad::git::identities::project::heads::default_branch_head` to
determine if there is an agreed on default branch commit and
`librad::git::identities::project::heads::set_default_branch` to set the
local `HEAD` ref where possible.

Signed-off-by: Alex Good <alex@memoryandthought.me>