lnk-identities: update path logic and set up include
The existing logic for checking out an identity in lnk-identities places
the checked out repository in `<selected path>/<identity name>` where
`selected path` is either the working directory or a specified
directory. This is not usually what people expect when checking out a
repository. Here we modify the logic so that if a directory is not
specified then we place the checked out repository in `$PWD/<identity
name>` but if the directory is specified then we place the checkout in
the specified directory directly.

While we're here we implement some missing logic to set the include path
in the newly created or updated repository.

Signed-off-by: Alex Good <alex@memoryandthought.me>