lnk-identities: improve track and untrack
The previous version of track and untrack were not updated to take
advantage of the latest `link-tracking` changes. As well as this, the
`inlcude::update` call would fail if the `urn` did not exist, making
it look like the tracking call failed at command line.

To remedy the former, make the `peer` optional to allow for default
tracking. Introduce `config` and `force` to be able to specify the
configuration used for the tracking and if it should forcefully
overwrite the configuration, respectively.

To remedy the later, check for `NotFound` error that can be returned by
the `include::update` call -- logging a warning.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>