lnk-identities: person/project delegates
A feature lacking from the `lnk identities` CLI is the ability to list
delegates of an identity. This *could* be derived from the output of
`tracked`. A dedicated subcommand is more convenient for this specific
output and is easier to remember.

Add a `delegates` subcommand to `person` and `project` that outputs
the delegates of the specified, verified identity. The output format
is a JSON list, each entry containing a JSON object:

    [ { "type": <type>, "payload": <payload> } ]

Currently, the type can be either "direct" or "indirect" -- the
payloads being a public key or an URN, respectively. This makes the
output of both commands uniform while also leaving room for new types
if they are ever added.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>