git-ext: fix to write
Remembering file modes and entering them into a text editor is error
prone when a compiler does not check you have done it
correctly. Unfortunately, there was a missing `0` for the file mode
for writing a tree entry.

Use the `FileMode` enum instead and its instance for `Into<i32>`.

The tracked_reference test that is using `quick_commit` was altered to
test that a tree with sub directories can be written.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>