clippy: run cargo update and switch to FuturesOrdered::push_back
The FuturesOrdered::push function was deprecated. Change to the
recommended FuturesOrdered::push_back.

To ensure everything compiles, run cargo update on the `bins`
workspace to ensure it is using futures-0.3.23.

Eerily, after updating the cli tests broke with missing parameters for
the args parsing tests. Fix this issue by adding the missing parameters.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>